Fix retained tell/gos communication locking in text from first person.
This is due to the way circlemud uses global buffers.  I used one buffer
that was used elsewhere and thus was overwriting the buffer with a fixed
version of the string instead of the variable.  For example, $n gossips,
'hi' will change into someone gossips 'hi' if the person gossiping is a
god. Since the $n is now someone, everyone after them in the
communicator chain will see someone instead of having it evaluate
correctly.